First commit of my Stm32f407 code.
This code has a working SPI SD Card. The I2C also works and the code calcuates the angle at which the accelerometer is held at. Note, the
Bias present on the X,Y & Z axies is not removed yet. The compass and gyroscope talks back to the micro as required.